refactor: Domain 패키지 분류
