+++ src/main/java/subway/application/LineService.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import subway.dao.LineDao;
 import subway.dao.entity.LineEntity;
-import subway.domain.Line;
-import subway.domain.LineColor;
-import subway.domain.LineName;
+import subway.domain.line.Line;
+import subway.domain.line.LineColor;
+import subway.domain.line.LineName;
 import subway.dto.line.LineCreateRequest;
 import subway.dto.line.LineResponse;
 import subway.dto.line.LineUpdateRequest;
