+++ src/main/java/subway/dao/SectionDao.java
@@ -3,7 +3,7 @@ package subway.dao;
 import java.util.List;
 import java.util.Optional;
 import subway.dao.entity.SectionEntity;
-import subway.domain.Section;
+import subway.domain.section.Section;
 
 public interface SectionDao {
