+++ src/main/java/subway/domain/line/LineName.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.line;
 
 import org.springframework.util.Assert;
