+++ src/main/java/subway/dto/station/StationResponse.java
@@ -1,6 +1,6 @@
 package subway.dto.station;
 
-import subway.domain.Station;
+import subway.domain.station.Station;
 
 public class StationResponse {
     private String stationName;
