+++ src/test/java/subway/dao/StationDaoTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import subway.domain.Station;
+import subway.domain.station.Station;
 
 @JdbcTest
 class StationDaoTest {
