+++ src/test/java/subway/domain/line/LineColorTest.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.line;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
