+++ src/test/java/subway/domain/section/DistanceTest.java
@@ -1,6 +1,7 @@
-package subway.domain;
+package subway.domain.section;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
