+++ src/test/java/subway/domain/section/SectionsTest.java
@@ -1,10 +1,11 @@
-package subway.domain;
+package subway.domain.section;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import subway.domain.station.Station;
 
 class SectionsTest {
